Moved code to corresponding locations for LHD functionalities
