+++ scripts/network_model.py
@@ -20,8 +20,6 @@ from scipy.sparse import csr_matrix
 from scipy.stats import truncnorm
 from typing import TypedDict, List, Dict, Any, Optional, Callable
 from collections import defaultdict
-from dataclasses import dataclass
-import uuid
 import warnings
 
 import matplotlib.pyplot as plt
@@ -29,7 +27,12 @@ from matplotlib.patches import Patch
 import matplotlib.animation as animation
 
 
+
 from scripts.synth_data_processing import build_edge_list, build_individual_lookup
+
+from new_scripts.recorder.recorder import ExposureEventRecorder
+from new_scripts.lhd.lhd import LocalHealthDepartment
+
 
 class ModelParameters(TypedDict):
     # Epi Params
@@ -1580,798 +1583,6 @@ class NetworkModel:
 
         return
 
-
-#------ Exposure Event Tracking --------
-class ExposureEventRecorder:
-    """
-    Object to record exposure events for a single timestep or across timesteps if reset. Stores:
-    - exposure event metadata arrays(time, source, type_id, start, length)
-    - nodes (concatenated node ids)
-    - infections (concatenated bool that align with nodes)
-    Methods:
-    - append_event(time, source, type_id, nodes_arr, infected_mask)
-    - reset() to reuse the recorder
-    - snapshot_compact(copy = True) -> dict of numpy arrays (sliced to used lengths)
-    - to_dataframe(id_to_ct) -> pandas DF with nodes and infected as arrays
-    """
-    def __init__(self, init_event_cap = 1024, init_node_cap = 4096):
-        self.event_cap = int(init_event_cap) 
-        self.node_cap = int(init_node_cap)
-        self._alloc_arrays()
-        self.reset()
-
-    def _alloc_arrays(self):
-        self.event_time = np.empty(self.event_cap, dtype = np.int32)
-        self.event_source = np.empty(self.event_cap, dtype = np.int32)
-        self.event_type = np.empty(self.event_cap, dtype = np.int16)
-        self.event_nodes_start = np.empty(self.event_cap, dtype = np.int64)
-        self.event_nodes_len = np.empty(self.event_cap, dtype = np.int32)
-        self.nodes = np.empty(self.node_cap, dtype = np.int32)
-        self.infections = np.empty(self.node_cap, dtype = np.bool_)
-
-    def reset(self):
-        self.n_events = 0
-        self.n_nodes = 0
-
-    def _grow_events(self, min_extra = 1):
-        if self.n_events + min_extra <= self.event_cap:
-            return
-        newcap = max(self.event_cap * 2, self.n_events + min_extra)
-        def grow(arr):
-            new = np.empty(newcap, dtype = arr.dtype)
-            new[: self.n_events] = arr[:self.n_events]
-            return new
-        
-        self.event_time = grow(self.event_time)
-        self.event_source = grow(self.event_source)
-        self.event_type = grow(self.event_type)
-        self.event_nodes_start = grow(self.event_nodes_start)
-        self.event_nodes_len = grow(self.event_nodes_len)
-        self.event_cap = newcap
-
-    def _grow_nodes(self, min_extra = 1):
-        if self.n_nodes + min_extra <= self.node_cap:
-            return
-        newcap = max(self.node_cap * 2, self.n_nodes + min_extra)
-        new_nodes = np.empty(newcap, dtype = np.int32)
-        new_nodes[: self.n_nodes] = self.nodes[: self.n_nodes]
-        new_inf = np.empty(newcap, dtype = np.bool_)
-        new_inf[: self.n_nodes] = self.infections[: self.n_nodes]
-        self.nodes = new_nodes
-        self.infections = new_inf
-        self.node_cap = newcap
-
-    def append_event(self, time: int, source: int, type_id: int, nodes_arr: np.ndarray, infected_mask: np.ndarray):
-        nodes_arr = np.asarray(nodes_arr, dtype = np.int32)
-        infected_mask = np.asarray(infected_mask, dtype = np.bool_)
-        if not nodes_arr.shape[0] == infected_mask.shape[0]:
-            raise ValueError("Recorder event has mismatched arrays")
-
-        L = nodes_arr.shape[0]
-
-        self._grow_events(1)
-        if L:
-            self._grow_nodes(L)
-            start = self.n_nodes
-            self.nodes[start:start + L] = nodes_arr
-            self.infections[start:start+L] = infected_mask
-        else:
-            start = self.n_nodes
-        
-        #gather metadata
-        i = self.n_events
-        self.event_time[i] = np.int32(time)
-        self.event_source[i] = np.int32(source)
-        self.event_type[i] = np.int16(type_id)
-        self.event_nodes_start[i] = np.int64(start)
-        self.event_nodes_len[i] = np.int32(L)
-
-        self.n_nodes += L
-        self.n_events += 1
-
-    def snapshot_compact(self, copy:bool = True) -> Dict[str, np.ndarray]:
-        """
-        Returns a dict of np.arrays sliced to used lengths
-
-        Args:
-            copy (bool): If true, arrays are copied and saved even if recorder is reused
-        """
-        event = slice(0, self.n_events)
-        node = slice(0, self.n_nodes)
-        if copy:
-            return{
-                'event_time': self.event_time[event].copy(),
-                'event_source': self.event_source[event].copy(),
-                'event_type': self.event_type[event].copy(),
-                'event_nodes_start': self.event_nodes_start[event].copy(),
-                'event_nodes_len': self.event_nodes_len[event].copy(),
-                'nodes': self.nodes[node].copy(),
-                'infections': self.infections[node].copy(),
-            }
-        else:
-            return {
-                'event_time': self.event_time[event],
-                'event_source': self.event_source[event],
-                'event_type': self.event_type[event],
-                'event_nodes_start': self.event_nodes_start[event],
-                'event_nodes_len': self.event_nodes_len[event],
-                'nodes': self.nodes[node],
-                'infections': self.infections[node],
-            }
-
-    def to_dataframe(self, id_to_ct: Dict[int, str]) -> pd.DataFrame:
-        """
-        Convert compact snapshot to a pandas dataframe. Computationally expensive
-
-        Args:
-            id_to_ct (Dict[int, str]): index to contact type mapping
-
-        """
-        rows = []
-        for i in range(self.n_events):
-            s = int(self.event_nodes_start[i])
-            L = int(self.event_nodes_len[i])
-            nodes = self.nodes[s:s+L].copy()
-            infs = self.infections[s:s+L].copy()
-            n_infected = sum(infs)
-            rows.append((
-                int(self.event_time[i]),
-                int(self.event_source[i]),
-                id_to_ct[int(self.event_type[i])],
-                nodes,
-                infs,
-                L,
-                n_infected
-               
-            ))
-        df = pd.DataFrame(rows, columns = ['time','source','contact_type','nodes','infected', 'n_exposed', 'n_infected'])
-        return(df)
-
-
-#------LHD Stuff --------
-
-
-####### Actions - LHD Interacting with the model
-
-#Token produced by action.apply and stored in expiry schedule
-@dataclass
-class ActionToken:
-    action_id: str
-    action_type: str #e.g. "calling", "quarantining", etc.
-    contact_type: Optional[str] #e.g. 'cas', 'hh' or None
-    nodes: np.ndarray #integer array of nodes affected by action
-    factor: Optional[np.ndarray] #factor applied to contact reductions or None
-    reversible: bool = True #if action can be undone
-    meta: Optional[Dict[str, Any]] = None #any relevant metadata
-
-class ActionBase:
-    """
-    Abstract action. Subclasses implement apply() to alter the model, returning ActionToken(s)
-    """
-    def __init__(self, action_type: str, duration:int = 0, kind: Optional[str] = None):
-        self.id = uuid.uuid4().hex
-        self.action_type = action_type
-        self.duration = int(duration)
-        self.kind = kind or action_type #more descriptive, sub-type label
-        self.reversible = True #subclasses can override
-
-    def apply(self, model: NetworkModel, current_time: int) -> List[ActionToken]:
-        """
-        Apply action to the model, return a list of ActionTokens describing state changes
-        """
-        raise NotImplementedError
-
-    def revert_token(self, model: NetworkModel, token: ActionToken) -> None:
-        """
-        Reverts a previously-applied token for a reversible action
-        """
-
-#Possible Actions
-class CallIndividualsAction(ActionBase):
-    """
-    Calls individuals to inform them of their exposure, applying a multiplicative reduction to their contact weight for school, workplace, and casual contacts
-
-    nodes: list of model node indices
-    contact_types: list or iterable of contact types ('cas', 'sch', etc.)
-    reduction: fraction (0-1) representing reduction (to be mult by compliance)
-    duration: days individual asked to reduce contacts
-    call_cost: hours spent per call
-    min_factor: minimum allowed reduction (just to avoid divzero errors)
-    """
-    def __init__(self, nodes: np.ndarray, contact_types: List[str], reduction: float, duration: int, call_cost: float, min_factor: float = 1e-6):
-        super().__init__(action_type = "call", duration = duration, kind = "call_individuals")
-        self.nodes = np.asarray(nodes, dtype = np.int32)
-        self.contact_types = list(contact_types)
-        self.reduction = float(reduction)
-        self.call_cost = float(call_cost)
-        self.min_factor = float(min_factor)
-        self.reversible = True
-
-    def apply(self, model: NetworkModel, current_time: int) -> List[ActionToken]:
-        #per-node factor, scaled by compliance
-        comps = model.compliances[self.nodes] 
-        raw = 1.0 - (self.reduction * comps)
-        factor = np.clip(raw, self.min_factor, 1.0).astype(np.float32)
-
-        tokens: List[ActionToken] = []
-        for ct in self.contact_types:
-            #apply reductions
-            model.out_multiplier[ct][self.nodes] *= factor
-            model.in_multiplier[ct][self.nodes] *= factor
-
-            token = ActionToken(
-                action_id = self.id,
-                action_type = self.action_type,
-                contact_type = ct,
-                nodes = self.nodes.copy(),
-                factor = factor.copy(),
-                reversible = True,
-                meta = {'call_cost': self.call_cost}
-            )
-            tokens.append(token)
-        
-        return tokens
-
-    def revert_token(self, model: NetworkModel, token: ActionToken) -> None:
-        #undo the multiplicative factor by dividing out 
-        if token.factor is None:
-            return
-        f = token.factor
-        #safeguard against zeros
-        f_safe = np.maximum(f, self.min_factor)
-        model.out_multiplier[token.contact_type][token.nodes] /= f_safe
-        model.in_multiplier[token.contact_type][token.nodes] /= f_safe
-
-
-
-
-####### Algorithms - LHD taking info from the outbreak and making decisions
-class AlgorithmBase:
-    """
-    An interface for designing calling algorithms for the local health department.
-    
-Algorithms have one method, generate_candidates, which takes:
-    recorder_snapshot: 
-    a NetworkModel object 
-    the indices of discovered events
-    
-    and returns arrays:
-         nodes
-         relative priority prioritization
-         contact type
-         projected call duration (call cost)
-    """
-    def generate_candidates(self, recorder_snapshot: Dict[str, np.ndarray], model, discovered_event_ind) -> Dict[str, Any]:
-        """
-
-        Args:
-            Snapshot of exposure events in an outbreak
-        Returns:
-            Dict[str, Any]: Dict with keys:
-                'nodes' : np.ndarray of node ids
-                'priority' np.ndarray of float priorities aligned w/ nodes
-                'contact_types': array of ct (per candidate or single)
-                other metadata
-        """
-        raise NotImplementedError
-
-## Possible Algorithms
-class EqualPriority(AlgorithmBase):
-    """
-    Returns each susceptible individual as equal priority candidates 
-    """
-    def generate_candidates(self, recorder_snapshot:Dict[str, np.ndarray], model, discovered_event_ind) -> Dict[str, Any]:
-        #if no events discovered, return nobody
-        if len(discovered_event_ind) == 0:
-            return {
-                'nodes': np.empty(0, dtype = np.int32),
-                'priority': np.empty(0, dtype = np.float32),
-                'contact_types': np.empty(0, dtype = object),
-                'costs': np.empty(0, dtype = np.float32),
-                'params': []
-            }
-
-        nodes_list = []
-        prios_list = []
-        cts_list = []
-        costs_list = []
-
-        default_cost = float(model.params.get('lhd_default_call_duration', 0.083))
-        for event in discovered_event_ind:
-            s = int(recorder_snapshot['event_nodes_start'][event])
-            L = int(recorder_snapshot["event_nodes_len"][event])
-            if L == 0:
-                continue
-
-            nodes = np.asarray(recorder_snapshot['nodes'][s:s+L], dtype = np.int32)
-            nodes_list.append(nodes)
-            prios_list.append(np.ones(nodes.shape[0], dtype = np.float32))
-            ct_name = model.id_to_ct[int(recorder_snapshot['event_type'][event])] if 'event_type' in recorder_snapshot else None
-            cts_list.append(np.array([ct_name]*nodes.shape[0], dtype = object))
-            costs_list.append(np.full(nodes.shape[0], default_cost, dtype = np.float32))
-
-        if not nodes_list:
-            return {
-                'nodes': np.empty(0, dtype = np.int32),
-                'priority': np.empty(0, dtype = np.float32),
-                'contact_types': np.empty(0, dtype = object),
-                'costs': np.empty(0, dtype = np.float32),
-                'params': []
-            }
-        return { 
-            'nodes': np.concatenate(nodes_list),
-            'priority': np.concatenate(prios_list),
-            'contact_types': np.concatenate(cts_list),
-            'costs': np.concatenate(costs_list),
-            'params': None
-        }
-
-class RandomPriority(AlgorithmBase):
-    """
-    Assigns each exposed candidate a random priority in [0-1)
-    """
-    def generate_candidates(self, recorder_snapshot: Dict[str, np.ndarray], model, discovered_event_ind) -> Dict[str, Any]:
-        #if no events discovered, return nobody
-        if len(discovered_event_ind) == 0:
-            return {
-                'nodes': np.empty(0, dtype = np.int32),
-                'priority': np.empty(0, dtype = np.float32),
-                'contact_types': np.empty(0, dtype = object),
-                'costs': np.empty(0, dtype = np.float32),
-                'params': []
-            }
-
-        nodes_list = []
-        prios_list = []
-        cts_list = []
-        costs_list = []
-        default_cost = float(model.params.get('lhd_default_call_duration', 0.083))
-        for event in discovered_event_ind:
-            s = int(recorder_snapshot['event_nodes_start'][event])
-            L = int(recorder_snapshot["event_nodes_len"][event])
-            if L == 0:
-                continue
-
-            nodes = np.asarray(recorder_snapshot['nodes'][s:s+L], dtype = np.int32)
-            nodes_list.append(nodes)
-
-            #random priority using model.rng
-            prios_list.append(model.rng.random(size = nodes.shape[0]).astype(np.float32))
-            ct_name = model.id_to_ct[int(recorder_snapshot['event_type'][event])] if 'event_type' in recorder_snapshot else None
-            cts_list.append(np.array([ct_name] * nodes.shape[0], dtype = object))
-            costs_list.append(np.full(nodes.shape[0], default_cost, dtype = np.float32))
-
-        if not nodes_list:
-            return {
-                'nodes': np.empty(0, dtype = np.int32),
-                'priority': np.empty(0, dtype = np.float32),
-                'contact_types': np.empty(0, dtype = object),
-                'costs': np.empty(0, dtype = np.float32),
-                'params': []
-            }
-
-        return { 
-            'nodes': np.concatenate(nodes_list),
-            'priority': np.concatenate(prios_list),
-            'contact_types': np.concatenate(cts_list),
-            'costs': np.concatenate(costs_list),
-            'params': None
-            }
-
-class PrioritizeElders(AlgorithmBase):
-    """
-    Priority boost for exposed nodes aged 65+. Priority = base + boost
-    Assume elders take longer to speak with on the phone 
-    """
-    def __init__(self, base_priority: float = 1.0, elder_boost: float = 4.0, elder_cost: float = 0.1 ):
-        self.base_priority = float(base_priority)
-        self.elder_boost = float(elder_boost)
-        self.elder_cost = float(elder_cost)
-
-    def generate_candidates(self, recorder_snapshot: Dict[str, np.ndarray], model, discovered_event_ind) -> Dict[str, Any]:
-        #if no events discovered, return nobody
-        if len(discovered_event_ind) == 0:
-            return {
-                'nodes': np.empty(0, dtype = np.int32),
-                'priority': np.empty(0, dtype = np.float32),
-                'contact_types': np.empty(0, dtype = object),
-                'costs': np.empty(0, dtype = np.float32),
-                'params': []
-            }
-
-        nodes_list = []
-        prios_list = []
-        cts_list = []
-        costs_list = []
-
-        default_cost = float(model.params.get('lhd_default_call_duration', 0.083))
-        for event in discovered_event_ind:
-            s = int(recorder_snapshot['event_nodes_start'][event])
-            L = int(recorder_snapshot["event_nodes_len"][event])
-            if L == 0:
-                continue
-
-            nodes = np.asarray(recorder_snapshot['nodes'][s:s+L], dtype = np.int32)
-            nodes_list.append(nodes)
-
-            ages_targets = model.ages[nodes]
-            pr = np.where(ages_targets >= 65, self.base_priority + self.elder_boost, self.base_priority).astype(np.float32)
-            prios_list.append(pr)
-
-            ct_name = model.id_to_ct[int(recorder_snapshot['event_type'][event])] if 'event_type' in recorder_snapshot else None
-            cts_list.append(np.array([ct_name] * nodes.shape[0], dtype = object))
-
-            costs = np.where(ages_targets >= 65, self.elder_cost, default_cost).astype(np.float32)
-            costs_list.append(costs)
-
-        if not nodes_list:
-            return {
-                'nodes': np.empty(0, dtype = np.int32),
-                'priority': np.empty(0, dtype = np.float32),
-                'contact_types': np.empty(0, dtype = object),
-                'costs': np.empty(0, dtype = np.float32),
-                'params': []
-            }
-
-        return { 
-            'nodes': np.concatenate(nodes_list),
-            'priority': np.concatenate(prios_list),
-            'contact_types': np.concatenate(cts_list),
-            'costs': np.concatenate(costs_list),
-            'params': None
-            }
-
-####### LHD Class - Local Health Department agent that interacts with outbreak
-class LocalHealthDepartment:
-    def __init__(
-        self, 
-        model: NetworkModel, 
-        rng = None, 
-        discovery_prob: float = None,  
-        employees: int = None, 
-        workday_hrs: float = None,
-        register_defaults: bool = True,
-        algorithm_map: Optional[Dict[str, object]] = None,
-        action_factory_map: Optional[Dict[str, Callable[..., ActionBase]]] = None
-    ):
-    #LHD settings
-        self.model = model
-        self.rng = rng if rng is not None else getattr(model, "rng", np.random.default_rng())
-        self.discovery_prob = discovery_prob if discovery_prob is not None else self.model.params["lhd_discovery_prob"]
-
-    #LHD Capacity
-        self.employees = employees if employees is not None else self.model.params["lhd_employees"]
-        self.hours_per_employee = float(workday_hrs) if workday_hrs is not None else self.model.params["lhd_workday_hrs"]
-        self.daily_personhours = float(self.employees * self.hours_per_employee)
-
-    #Algorithm -> algorithm instance
-        self.algorithms: Dict[str, AlgorithmBase] = {}
-    #action factories: action_type -> callable to return ActionBase
-        self.action_factories: Dict[str, Callable[..., ActionBase]] = {}
-
-
-    #trackers for action objects and token counts
-        self.expiry: Dict[int, List[ActionToken]] = {}
-        self.action_log: List[Dict[str, Any]] = []
-        # action id -> action instance
-        self._active_actions: Dict[str, ActionBase] = {}
-        #action id -> number outstanding tokens
-        self._action_token_counts: Dict[str, int] = {}
-
-        self.min_factor = 1e-6 #to prevent div 0 errors
-        self.min_candidate_cost = 1e-4
-
-    #Default action params
-        self.default_int_reduction = model.params.get("lhd_default_int_reduction", 0.8)
-        self.default_int_duration = model.params.get("lhd_default_int_duration", 7)
-        self.default_call_cost = model.params.get("lhd_default_call_duration", 0.083)
-
-
-        #Register default actions if requested:
-        if register_defaults:
-            self.register_algorithm('call', RandomPriority())
-
-            def default_call_factory(nodes, contact_type, prio, cost, params = None):
-                return CallIndividualsAction(
-                nodes = nodes,
-            contact_types = [contact_type] if contact_type is not None else
-            ['cas', 'sch', 'wp'],
-            reduction = params.get('reduction', self.default_int_reduction) if params else self.default_int_reduction,
-            duration = int(params.get('duration', self.default_int_duration)) if params else self.default_int_duration,
-            call_cost = float(cost) if cost is not None else self.default_call_cost,
-            min_factor = self.min_factor
-        )
-            self.register_action_factory('call', default_call_factory)
-
-        #Register mappings provided by call
-        if algorithm_map:
-            for atype, alg in algorithm_map.items():
-                self.register_algorithm(atype, alg, overwrite = True)
-        if action_factory_map:
-            for atype, factory in action_factory_map.items():
-                self.register_action_factory(atype, factory, overwrite = True)
-            
-
-
-    ##registration helpers
-    # map algorithms action_type -> algorithm
-    def register_algorithm(self, action_type: str, algorithm: AlgorithmBase, overwrite: bool = False) -> None:
-        """
-        Assign each action with an algorithm that is used to decide who that action should be done to. 
-        """
-        if action_type in self.algorithms and not overwrite:
-            raise ValueError(f"Algorithm already registered for action '{action_type}'. Only one allowed.")
-        if action_type in self.algorithms and overwrite:
-            warnings.warn(f"Overwriting existing algorithm for action '{action_type}'")
-        self.algorithms[action_type] = algorithm
-
-    #map action_type -> factory to create action objects
-    #expects 
-    def register_action_factory(self, action_type: str, factory: Callable[..., ActionBase], overwrite: bool = False) -> None:
-        #factory signature expected: (nodes, contact_type, prio, cost, params) -> ActionBase
-        if action_type in self.action_factories and not overwrite:
-            raise ValueError(f"Action factory already registered for action '{action_type}'. Only one allowed.")
-        if action_type in self.action_factories and overwrite:
-            warnings.warn(f"Overwriting existing action factory for action '{action_type}'")
-        self.action_factories[action_type] = factory
-
-    def discover_exposures(self, recorder_snapshot):
-        """
-        Given recorder snapshot dict, select which events the LHD discovers
-        Random based on LHD discovery probability
-        """
-        n_events = recorder_snapshot['event_time'].shape[0]
-        if n_events == 0:
-            return np.empty(0, dtype = int)
-        
-        #bernoulli sample each event
-        mask = self.rng.random(n_events) < self.discovery_prob
-        return np.where(mask)[0]
-
-    def gather_candidates(self, recorder_snapshot, discovered_event_ind):
-        """
-        Returns flattened candidate arrays with parallel arrays:
-        action_types (st array), nodes (int), priority (float), 
-        contact_types (object), costs (float)
-        """
-        
-        #lists to fill
-        action_types_list = []
-        nodes_list = []
-        prios_list = []
-        cts_list = []
-        costs_list = []
-        params_list = []
-
-        #use algorithms to generate action candidates
-        for action_type, algo in self.algorithms.items():
-            out = algo.generate_candidates(recorder_snapshot, self.model, discovered_event_ind) or {}
-            nodes = np.asarray(out.get('nodes', np.empty(0, dtype=np.int32)), dtype=np.int32)
-            prios = np.asarray(out.get('priority', np.ones(nodes.shape[0], dtype=np.float32)), dtype=np.float32)
-
-            #raw cts can be a single str or array of size len(nodes)
-            raw_cts = out.get('contact_types', None)
-            if raw_cts is None:
-                cts = np.array([None] * nodes.shape[0], dtype = object)
-            elif isinstance(raw_cts, (str, bytes)):
-                cts =  np.array([raw_cts] * nodes.shape[0], dtype = object)
-            else:
-                cts = np.asarray(raw_cts, dtype = object)
-                if cts.shape[0] != nodes.shape[0]:
-                    if cts.size == 1:
-                        cts = np.repeat(cts[0], nodes.shape[0]).astype(object)
-                    else:
-                        raise ValueError("contact_types must be scalar or match nodes length")
-            
-            #costs must be a single float or array of size len(nodes)
-            costs = out.get('costs', None)
-            if costs is None:
-                costs = np.full(nodes.shape[0], np.nan, dtype = np.float32)
-            else:
-                costs = np.asarray(costs, dtype = object)
-                if costs.shape[0] != nodes.shape[0]:
-                    if costs.size == 1:
-                        costs = np.repeat(costs[0], nodes.shape[0]).astype(object)
-                    else:
-                        raise ValueError("costs must be scalar or match nodes length")
-
-            params = out.get('params', None)
-            #for each action type, deduplicate candidates
-            if nodes.size > 0:
-                unique_nodes, inverse = np.unique(nodes, return_inverse = True)
-                best_prios = np.full(unique_nodes.shape[0], -np.inf, dtype = np.float32)
-                best_cts = np.empty(unique_nodes.shape[0], dtype = object)
-                best_costs = np.full(unique_nodes.shape[0], np.inf, dtype = np.float32)
-                best_params = [None] * unique_nodes.shape[0]
-
-                for occ in range(nodes.shape[0]):
-                    uid = inverse[occ]
-                    p = float(prios[occ])
-                    c = float(costs[occ])
-                    #keep highest priority, or lowest cost if tie
-                    if p > best_prios[uid] or (p==best_prios[uid] and c < best_costs[uid]):
-                        best_prios[uid] = p
-                        best_cts[uid] = cts[occ]
-                        best_costs[uid] = c
-                        if params is not None:
-                            #params can be per-occurrance or scalar
-                            try:
-                                best_params[uid] = params[occ]
-                            except Exception:
-                                best_params[uid] = params
-                    
-                #append best occurrences to global lists with action_type label
-                for i, u in enumerate(unique_nodes):
-                    action_types_list.append(action_type)
-                    nodes_list.append(int(u))
-                    prios_list.append(best_prios[i])
-                    cts_list.append(best_cts[i])
-                    costs_list.append(best_costs[i])
-                    params_list.append(best_params[i])
-
-        #if no nodes to gather, return empties
-        if not nodes_list:
-            return(
-                np.empty(0, dtype = object),
-                np.empty(0, dtype = np.int32), 
-                np.empty(0, dtype = np.float32),
-                np.empty(0, dtype = object),
-                np.empty(0, dtype = np.float32), 
-                []
-            )
-
-        #else, gather results and return
-        action_types_arr = np.array(action_types_list, dtype = object)
-        nodes_arr = np.array(nodes_list, dtype = np.int32)
-        prios_arr = np.array(prios_list, dtype = np.float32)
-        contact_types_arr = np.array(cts_list, dtype = object)
-        costs_arr = np.array(costs_list, dtype = np.float32)
-        params_arr = params_list
-
-        return action_types_arr, nodes_arr, prios_arr, contact_types_arr, costs_arr, params_arr
-
-    def schedule_action(self, action: ActionBase, current_time: int, resource_cost: float):
-        """
-        Apply action and schedule tokens for expiry if duration > 0.
-        Registers Action instance for process_expirations to call reversion
-        """
-        #Apply all actions, get a list of actions performed
-        tokens = action.apply(self.model, current_time)
-
-        for t in tokens:
-            if getattr(t, "action_id", None) != action.id:
-                raise ValueError(f"Token.action_id {getattr(t, 'action_id', None)} does not match action.id {action.id}")
-
-        #partition to reversible and nonreversible
-        reversible_tokens = [t for t in tokens if getattr(t, "reversible", True)]
-        nonreversible_tokens = [t for t in tokens if not getattr(t, "reversible", True)]
-
-        #if there are reversible tokens and duration, schedule
-
-
-        #register reversible tokens only, schedule expiry
-        if reversible_tokens and (action.duration and action.duration > 0):
-            expiry_time = int(current_time + action.duration)
-            self.expiry.setdefault(expiry_time, []).extend(reversible_tokens)
-
-            self._active_actions[action.id] = action
-            self._action_token_counts[action.id] = self._action_token_counts.get(action.id, 0) + len(reversible_tokens)
-
-        #if duration > 0, but not reversible, warning as nothing to revert
-        if action.duration and action.duration > 0 and not reversible_tokens:
-            warnings.warn(f"Action {action.id} has duration but produced no reversible tokens, will not be automatically reverted")
-
-        #log metadata
-        self.action_log.append({
-            'time': int(current_time),
-            'action_id': action.id,
-            'action_type': action.action_type,
-            'kind': getattr(action, "kind", action.action_type),
-            'nodes_count': int(getattr(action, "nodes", np.empty(0)).size),
-            'hours_used': float(resource_cost),
-            'duration': int(action.duration),
-            'reversible_tokens': len(reversible_tokens),
-            'nonreversible_tokens': len(nonreversible_tokens)
-        })
-
-    def process_expirations(self, current_time):
-        """
-        Revert ActionTokens scheduled for current_time using action reversion methods, and remove from active actions
-        """
-        tokens_due = self.expiry.pop(int(current_time), [])
-        for token in tokens_due:
-            action = self._active_actions.get(token.action_id)
-            if action is not None:
-                #delegate reversion to action's method
-                try:
-                    action.revert_token(self.model, token)
-                except Exception as exc:
-                    warnings.warn(f"Action.revert_token failed for action {token.action_id}: {exc}")
-
-                #decrement token counters and clean mappings 
-                self._action_token_counts[token.action_id] -= 1
-                if self._action_token_counts[token.action_id] <= 0:
-                    #if no more tokens on this action, remove from active 
-                    del self._action_token_counts[token.action_id]
-                    del self._active_actions[token.action_id]
-            else:
-                warnings.warn(f"No registered action object for token.action_id {token.action_id}")
-
-    def step(self, current_time: int, recorder_snapshot: Dict[str, np.ndarray]):
-        """
-        One step for the LHD where it discovers events, builds candidates, selects calls, and applies interventions
-        """
-
-        #1 expire old interventions
-        self.process_expirations(current_time)
-
-        #2 discover new events
-        discovered_event_ind = self.discover_exposures(recorder_snapshot)
-
-        #3 gather action candidates through algorithms
-        (action_types_arr, nodes_arr, prios_arr, contact_types_arr, costs_arr, params_arr) = self.gather_candidates(recorder_snapshot, discovered_event_ind)
-        if nodes_arr.size == 0:
-            return
-
-        #4 select actions maximizing value/hour (prio / cost)
-        costs_arr = np.maximum(costs_arr, self.min_candidate_cost)
-        value_per_hour = prios_arr/costs_arr
-        order = np.argsort(-value_per_hour)
-
-        #allocate actions by value until hours are exhausted 
-        hours_available = float(self.daily_personhours)
-        hours_spent = 0.0
-        selected_indices = []
-        for ind in order:
-            c = float(costs_arr[ind])
-            if hours_spent + c <= hours_available:
-                hours_spent += c
-                selected_indices.append(ind)
-            else:
-                continue
-
-        #5 group selected actions by action_type, contact_type and schedule
-        grouped = defaultdict(list)
-        grouped_costs = defaultdict(float)
-        grouped_params = defaultdict(list)
-        for ind in selected_indices:
-            atype = action_types_arr[ind]
-            ctype = contact_types_arr[ind]
-            key = (atype, ctype)
-            grouped[key].append(int(nodes_arr[ind]))
-            grouped_costs[key] += float(costs_arr[ind])
-            grouped_params[key].append(params_arr[ind] if params_arr is not None else None)
-
-        #create an action for each group and schedule
-        for (atype, ctype), nodes in grouped.items():
-            factory = self.action_factories.get(atype)
-            if factory is None:
-                #skip if no action factory registered
-                continue
-            #choose params and pass merged dict or None
-            params_list_group = grouped_params[(atype, ctype)]
-            merged_params = None
-            for p in params_list_group:
-                if isinstance(p, dict):
-                    merged_params = merged_params or {}
-                    merged_params.update(p)
-
-            #create action instance 
-            #use sum of costs or average priority for group
-            group_cost = grouped_costs[(atype, ctype)]
-            group_prio = float(np.mean([prios_arr[ind] for ind in selected_indices if action_types_arr[ind] == atype and contact_types_arr[ind] == ctype]))
-            action = factory(np.asarray(nodes, dtype = np.int32), ctype, group_prio, group_cost, merged_params)
-            self.schedule_action(action, current_time, resource_cost = group_cost)
-
-        return
-    def reset_for_run(self):
-        """
-        Reset LHD state for new model run 
-        """
-        self.expiry = {}
-        self.action_log = []
-        self._active_actions = {}
-        self._action_token_counts = {}
 
 
 #Test run on population N = 2186
